separated spec files for cta
