+++ cypress/integration/v2/blocks/cta/desktop-advanced.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { desktopAdvanced } from '.'
+
+describe( 'Call To Action Block ( Desktop Advanced )', registerTests( [ desktopAdvanced ] ) )
