+++ cypress/integration/v2/blocks/cta/desktop-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { desktopStyle } from '.'
+
+describe( 'Call To Action Block ( Desktop Style )', registerTests( [ desktopStyle ] ) )
