+++ cypress/integration/v2/blocks/cta/tablet-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { tabletStyle } from '.'
+
+describe( 'Call To Action Block ( Tablet Style )', registerTests( [ tabletStyle ] ) )
